requirements.txt
alembic==1.15.2
blinker==1.9.0
cachelib==0.13.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.0
cryptography==45.0.2
Flask==3.1.1
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
msal==1.32.3
msgspec==0.19.0
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.10.1
python-dotenv==1.1.0
requests==2.32.3
SQLAlchemy==2.0.41
typing_extensions==4.13.2
urllib3==2.4.0
Werkzeug==2.2.2
WTForms==3.2.1
